move spec validation into seperate file
